+++ CHANGELOG.md
@@ -1,9 +1,20 @@
 ## Changelog
 
+### [1.1.0] - 2024-08-12
+
+Added
+
+- Now can set initialData and initialSource
+
+Fixed
+
+- lastFetchTime being properly set
 
 ### [1.0.1] - 2024-08-12
 
-Fixed source calling .subscribe immediately
+Fixed
+
+- source calling .subscribe immediately
 
 ### [1.0.0] - 2024-08-12